pyproject.toml
[tool.poetry]
name = "serverless-echo"
version = "0.1.0"
description = ""
authors = ["Max Friedrich <hallo@maxfriedrich.de>"]
readme = "README.md"
packages = [{include = "echo"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
